Add SelectableListItem widget
